requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
python-dotenv==1.0.1
pydantic==2.11.7
qrcode[pil]==7.4.2
python-multipart==0.0.20
pypdf==5.9.0
eval_type_backport==0.2.2
